requirements.txt
scapy>=2.4.5
pysctp>=0.7.2
pyyaml>=6.0
colorama>=0.4.4
pytest>=7.0.0